feat: add compact hero
